requirements.txt
# 🌞 HelioBio-Social v1.0.0 - Dependencias Espirituales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
pydantic==2.5.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.13.0
plotly==5.17.0
google-cloud-aiplatform==1.38.0
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3
watchfiles==0.21.0
jinja2==3.1.2
